# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Cервер на Golang с OpenAPI документацией.",
        "title": "Shop API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Maksim",
            "url": "https://github.com/hhullen",
            "email": "hhullen@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/client": {
            "post": {
                "description": "Добавление клиента. Если клиент существует вернется uid существующего клиента.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "summary": "Добавление клиента",
                "parameters": [
                    {
                        "description": "Информация о клиенте",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/datastruct.AddClientRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/datastruct.AddClientResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/datastruct.Status"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/datastruct.Status"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет клиента по его uid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "summary": "Удаление клиента",
                "parameters": [
                    {
                        "description": "uid клиента",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/datastruct.DeleteClientRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/datastruct.DeleteClientResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/datastruct.DeleteClientResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/datastruct.Status"
                        }
                    }
                }
            }
        },
        "/client/address": {
            "patch": {
                "description": "Обновляет адрес клиента по его uid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "summary": "Обновляет адрес клиента",
                "parameters": [
                    {
                        "description": "uid и адрес",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/datastruct.PatchClientAddressRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/datastruct.PatchClientAddressResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/datastruct.PatchClientAddressResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/datastruct.Status"
                        }
                    }
                }
            }
        },
        "/clients": {
            "get": {
                "description": "Возвращает список клиентов. Если offset limit равны 0, вернет список всех клиентов",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "summary": "Возвращает список клиентов",
                "parameters": [
                    {
                        "type": "string",
                        "example": "0",
                        "description": "offset",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "10",
                        "description": "limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "true",
                        "description": "avoid_cache",
                        "name": "avoid_cache",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/datastruct.GetClientsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/datastruct.Status"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/datastruct.Status"
                        }
                    }
                }
            }
        },
        "/clients/named": {
            "get": {
                "description": "Возвращает клиентов по имени и фамилии",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "summary": "Возвращает клиентов по имени и фамилии",
                "parameters": [
                    {
                        "type": "string",
                        "example": "Vasilisa",
                        "description": "client_name",
                        "name": "client_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "Kadyk",
                        "description": "client_surname",
                        "name": "client_surname",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "true",
                        "description": "avoid_cache",
                        "name": "avoid_cache",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/datastruct.GetClientsByNameResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/datastruct.Status"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/datastruct.Status"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "datastruct.AddClientRequest": {
            "type": "object",
            "required": [
                "address",
                "birthday",
                "client_name",
                "client_surname",
                "gender",
                "registration_date"
            ],
            "properties": {
                "address": {
                    "$ref": "#/definitions/datastruct.Address"
                },
                "avoid_cache": {
                    "type": "boolean",
                    "example": true
                },
                "birthday": {
                    "type": "string",
                    "example": "10.12.2011"
                },
                "client_name": {
                    "type": "string",
                    "example": "Vasilisa"
                },
                "client_surname": {
                    "type": "string",
                    "example": "Kadyk"
                },
                "gender": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/datastruct.Gender"
                        }
                    ],
                    "example": "female"
                },
                "registration_date": {
                    "type": "string",
                    "example": "30/01/2026"
                },
                "uid": {
                    "type": "string",
                    "example": "4988150e-1c82-490f-8c07-ee74ace2dd14"
                }
            }
        },
        "datastruct.AddClientResponse": {
            "type": "object",
            "properties": {
                "cached": {
                    "type": "boolean",
                    "example": false
                },
                "status": {
                    "type": "string",
                    "example": "status message"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
...
        "datastruct.Address": {
            "type": "object",
            "required": [
                "city",
                "country",
                "street"
            ],
            "properties": {
                "city": {
                    "type": "string",
                    "example": "Seattle"
                },
                "country": {
                    "type": "string",
                    "example": "USA"
                },
                "street": {
                    "type": "string",
                    "example": "12th Ave E"
                }
            }
        },
        "datastruct.Client": {
            "type": "object",
            "required": [
                "address",
                "birthday",
                "client_name",
                "client_surname",
                "gender",
                "registration_date"
            ],
            "properties": {
                "address": {
                    "$ref": "#/definitions/datastruct.Address"
                },
                "birthday": {
                    "type": "string",
                    "example": "10.12.2011"
                },
                "client_name": {
                    "type": "string",
                    "example": "Vasilisa"
                },
                "client_surname": {
                    "type": "string",
                    "example": "Kadyk"
                },
                "gender": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/datastruct.Gender"
                        }
                    ],
                    "example": "female"
                },
                "registration_date": {
                    "type": "string",
                    "example": "30/01/2026"
                },
                "uid": {
                    "type": "string",
                    "example": "4988150e-1c82-490f-8c07-ee74ace2dd14"
                }
            }
        },
...
        "datastruct.DeleteClientRequest": {
            "type": "object",
            "required": [
                "uid"
            ],
            "properties": {
                "avoid_cache": {
                    "type": "boolean",
                    "example": true
                },
                "uid": {
                    "type": "string",
                    "example": "4988150e-1c82-490f-8c07-ee74ace2dd14"
                }
            }
        },
        "datastruct.DeleteClientResponse": {
            "type": "object",
            "properties": {
                "cached": {
                    "type": "boolean",
                    "example": false
                },
                "status": {
                    "type": "string",
                    "example": "status message"
                }
            }
        },
...
        "datastruct.Gender": {
            "type": "string",
            "enum": [
                "male",
                "female"
            ],
            "x-enum-varnames": [
                "Male",
                "Female"
            ]
        },
        "datastruct.GetClientsByNameResponse": {
            "type": "object",
            "properties": {
                "cached": {
                    "type": "boolean",
                    "example": false
                },
                "clients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/datastruct.Client"
                    }
                }
            }
        },
        "datastruct.GetClientsResponse": {
            "type": "object",
            "properties": {
                "cached": {
                    "type": "boolean",
                    "example": false
                },
                "clients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/datastruct.Client"
                    }
                }
            }
        },
...
        "datastruct.PatchClientAddressRequest": {
            "type": "object",
            "required": [
                "address",
                "uid"
            ],
            "properties": {
                "address": {
                    "$ref": "#/definitions/datastruct.Address"
                },
                "avoid_cache": {
                    "type": "boolean",
                    "example": true
                },
                "uid": {
                    "type": "string",
                    "example": "4988150e-1c82-490f-8c07-ee74ace2dd14"
                }
            }
        },
        "datastruct.PatchClientAddressResponse": {
            "type": "object",
            "properties": {
                "cached": {
                    "type": "boolean",
                    "example": false
                },
                "status": {
                    "type": "string",
                    "example": "status message"
                }
            }
        },
...
        "datastruct.Status": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "status message"
                }
            }
        },
...
    }
}